refactor Order Screens so all components live on NewOrderScreen and delete unneeded files.
